% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydra.R
\docType{data}
\name{ExamplePleSpecs}
\alias{ExamplePleSpecs}
\title{Example PLE json specification}
\format{
A character containing the json settings
}
\usage{
data(ExamplePleSpecs)
}
\description{
An example json specification object used by Hydra
}
\examples{
\donttest{data(ExamplePleSpecs)

hydrate(specifications = ExamplePleSpecs, 
outputFolder = file.path(getwd(), 'example'), 
skeletonFileName = NULL, 
packageName = NULL))}
}
\keyword{datasets}
