% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestHelpers.R
\name{getTestEstimationSpecification}
\alias{getTestEstimationSpecification}
\title{Helper function for obtaining a JSON specification for testing the estimation package}
\usage{
getTestEstimationSpecification()
}
\description{
Helper function for obtaining a JSON specification for testing the estimation package
}
