% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydra.R
\docType{package}
\name{Hydra-package}
\alias{Hydra}
\alias{Hydra-package}
\title{Hydra: Hydrating Package Skeletons into Executable R Study Packages}
\description{
An R package and Java library for hydrating package skeletons into executable R study packages based on specifications in JSON format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/Hydra}
  \item \url{https://github.com/OHDSI/Hydra}
  \item Report bugs at \url{https://github.com/OHDSI/Hydra/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Jenna Reps
  \item Anthony Sena
}

}
\keyword{internal}
