% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydratePackage.R
\name{hydrate}
\alias{hydrate}
\title{Hydrate a package skeleton according to study specifications}
\usage{
hydrate(
  specifications,
  outputFolder,
  skeletonFileName = NULL,
  packageName = NULL
)
}
\arguments{
\item{specifications}{A string containing JSON specifications for hydrating 
the study. Specifications will be generated by ATLAS.}

\item{outputFolder}{The folder where the hydrated package will be stored.}

\item{skeletonFileName}{(Mostly for debugging purposes:) Specify the location
of a package skeleton to hydrate. If not specified,
the appropriate skeleton will be selected from those
inside the Hydra package.}

\item{packageName}{The name of the R package to create. If provided, will
override the \code{packageName} variable in the 
specifications JSON.}
}
\description{
Hydrate a package skeleton according to study specifications
}
