% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OfflineStudyPackageExecution.R
\name{listSkeletons}
\alias{listSkeletons}
\title{List skeletons included in Hydra}
\usage{
listSkeletons()
}
\value{
A vector of skeleton names.
}
\description{
List skeletons included in Hydra
}
\examples{
listSkeletons()

}
