% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydratePackage.R
\name{loadSpecifications}
\alias{loadSpecifications}
\title{Load specifications from a JSON file}
\usage{
loadSpecifications(fileName)
}
\arguments{
\item{fileName}{Path to the JSON file}
}
\value{
A string representation of the JSON.
}
\description{
Load specifications from a JSON file
}
