% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OfflineStudyPackageExecution.R
\name{prepareForOfflineStudyPackageExecution}
\alias{prepareForOfflineStudyPackageExecution}
\title{Prepare system to run hydrated packages without further internet connection}
\usage{
prepareForOfflineStudyPackageExecution(
  installRpackages = TRUE,
  installJdbcDrivers = TRUE,
  skeletonVector = NULL
)
}
\arguments{
\item{installRpackages}{Install the R packages required by the skeletons?}

\item{installJdbcDrivers}{Install all JDBC drivers? Requires the DATABASECONNECTOR_JAR_FOLDER
environmental variable to be set.}

\item{skeletonVector}{list of skeletons to check, if null uses all in the Hydra package}
}
\value{
This function does not return anything. Instead, it installs all dependencies required
to run the hydrated skeletons. (Only those skeletons that use a renv.lock file)
}
\description{
Prepare system to run hydrated packages without further internet connection
}
\details{
Note that when \code{installJdbcDrivers = TRUE} this will only include the Jar drivers
supported by DatabaseConnector::downloadJdbcDrivers. Other drivers (like BigQuery) will
need to be downloaded manually and placed in the folder identified by the 
DATABASECONNECTOR_JAR_FOLDER environmental variable.
}
